requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
autopep8==1.5.4
beautifulsoup4==4.9.1
cement==3.0.4
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django @ file:///tmp/build/80754af9/django_1600355803933/work
django-extensions==3.0.9
djangorestframework==3.11.1
drf-yasg==1.17.1
httpie==2.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1594305108220/work
pycodestyle==2.6.0
Pygments==2.7.1
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
